Don't set CMAKE_WINDOWS_KITS_10_DIR from environment
This avoids situations where WindowsSdkDir is set in the environment but we want to select a specific version via the pre-existing mechanisms in thie toolchain.
It's better to require EWDK users to explicitly state what they want by setting CMAKE_WINDOWS_KITS_10_DIR via cmake cache vars on the commandline.